+++ R/ALASCA.R
@@ -19,6 +19,7 @@
 #' @param plot.figsize A vector containing `c(width,height,dpi)` (default: `c(120, 80, 300)`)
 #' @param plot.figunit Defaults to "mm"
 #' @param plot.filetype Which filetype you want to save the figure to (default: `png`)
+#' @param plot.palette List of colors, named by group
 #' @param keepTerms Additional terms to keep in the model matrix
 #' @param stratificationVector Vector of same length as `df` that specifies stratification groups during validation. Defaults to `NA`, where the group column is used.
 #' @param validateRegression Whether to validate regression predictions or not (only if `validate` is `TRUE`)
@@ -58,6 +59,7 @@ ALASCA <- function(df,
                    plot.figsize = c(12, 8, 300),
                    plot.figunit = "mm",
                    plot.filetype = "png",
+                   plot.palette = NA,
                    doDebug = FALSE,
                    nValFold = 7,
                    nValRuns = 50,
@@ -110,6 +112,7 @@ ALASCA <- function(df,
                    plot.figsize = plot.figsize,
                    plot.figunit = plot.figunit,
                    plot.filetype = plot.filetype,
+                   plot.palette = plot.palette,
                    minimizeObject = minimizeObject,
                    doDebug = doDebug,
                    nValFold = nValFold,
@@ -179,7 +182,7 @@ RMASCA <- function(...){
 #' @return String
 #' @export
 printVer <- function(object = FALSE, get = NA, print = TRUE){
-  ALASCA.version <- "0.0.0.97"
+  ALASCA.version <- "0.0.0.98"
   ALASCA.version.date <- "2022-01-04"
   if(is.list(object)){
     ALASCA.version <- object$ALASCA.version